file-bin: Support reading files larger than ~4Gb!
Some variables representing offsets in the file were uint32_t, so
limited to ~4Gb. Because of that, reading files larger than 4Go was not
supported.

Use off_t to fix that.

Priv-Id: d3d7eb10be6b1c5f98b070862c368bb89d6a86d4

+++ src/core/file-bin.c
@@ -261,7 +261,7 @@ static int _file_bin_get_next_record(file_bin_t *file, lstr_t *rec)
     off_t prev_off = file->cur;
     off_t rec_end_off;
     uint32_t sz;
-    uint32_t check_slot_hdr;
+    off_t check_slot_hdr;
     bool is_spanning = false;
 
     if (file_bin_is_finished(file)) {
